src/main/python/requirements.txt
# requirements.txt

blinker==1.6.3
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
grpcio==1.59.0
grpcio-tools==1.59.0
humanfriendly==10.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
numpy==1.26.0
onnx==1.14.1
onnxruntime==1.16.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
protobuf==4.24.4
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sympy==1.12
typing_extensions==4.8.0
Werkzeug==3.0.0
